Filled dockerfile as needed
The used image is python:3.13-alpine which is the latest image that
supports python out of the box.
Working directory set to /app - industry standard.
Copied and installed requirements.txt with pip so the necessary
packages are present for the application.
Copied application to working directory.
Set the command to the same as found in run.sh, which will automatically
start the application once the image is run.

Solves #1